Add Unicode support to key up/down events
The key up/down events kup(...) and kdown(...) now take StringRefs.

KeySym functionality has also been enhanced - it now allows the encoding of Unicode codepoints in addition to the traditional ASCII and special keys.

A greater emphasis has been placed on whether an object cares about key presses or simply text. The "field" object is obviously interested in text but buttons (in the guise of pushbuttons, menus, etc) are really concerned with a particular keystroke being used as an accelerator or mnemonic. Therefore, field objects make use of the p_string parameter of the kup/kdown message while buttons instead use KeySym.

Theoretically, this allows any Unicode codepoint to be used as an accelerator or mnemonic for menus and buttons, respectively.