MCConstant stores its string value as a StringRef
Fixed a very obscure and not script parser bug on the way... (literal tokens should not be looked up in the symbol tables). As an example, the following script would trigger the bug (which was only uncovered due to a change in the ScriptPoint internals):

if variable is "a" then

The literal "a" would be looked up in a certain symbol table (validation) where it would be found, causing the syntax to be treated as if it was instead:

if variable is a then

which is a syntax error as "then" is not a type. Other lines could trigger the bug but all were characterised by a literal being erroneously treated as an identifier.