Add MCStringCreateWithSysString and MCStringConvertToSysString
These functions are used on Linux to convert between ISO-8859-1 or UTF-16 encoded StringRefs used by LiveCode and whatever string encoding the Linux system expects (probably UTF-8 or ISO-8859-1 but it could be anything). Windows doesn't need this as its *W() functions always take UTF-16 and on OSX the BSD components should always use UTF-8.

Currently, these functions are only used for ARGV/ENVP strings but ideally should be used whenever a string is passed to a system interface (file APIs, for example).

String handling when spawning elevated processes is not yet correct - it currently assumes that the system encoding is UTF-8 when passing argument strings. This affects both the exec of gksu and the creation of the final elevated process.