Change hash function used for text measurement cache
Previously, the hash used by the ELF file for symbol lookups was used. This was not very suitable and led to frequent collisions; the average lookup time in the cache was linear in the number of occupied buckets instead of constant. It has been replaced with the 32-bit variety of the FNV-1a hash and performance is (nearly) constant regardless of table occupancy.