Bugfix 11285
On iOS7 devices, the [UITextView font] method returned nil if no text had yet been added to the control so attempts to get font size were ignored. A check has now been added for this condition and the system font of the appropriate size is used if a custom font has not been set.