libfoundation: Provide ATTRIBUTE_UNUSED macro.
In some places, it makes sense to define (and possibly set) a
variable, but not to use it.  In that case,
__attribute__((__unused__)) to the variable suppresses compiler
warnings, e.g.

    int var_name ATTRIBUTE_UNUSED;

The alternative is to add a line:

    (void) var_name;

which will also suppress the warnings.  However, ATTRIBUTE_UNUSED is
more easy to search for across the source code.