[[feature-stackdir]] Spec: Optional SHA-1s; property flags.
Some externally-stored properties are expected to be regularly
modified concurrently by multiple users -- for example, scripts.
Storing a SHA-1 of the file means that a merge conflict will occur for
every concurrent modification and will have negative workflow
consequences (e.g. it will prevent the use of pull requests on
Github).  We are therefore pursuing a different approach to solving
the problem of concurrent deletion and modification.

It is useful to store optional flag bits along with property
descriptors.  These have a number of applications.