libfoundation: Unify record field-counting algorithms.
It doesn't make sense to have two implementations of the "count all
fields in record type and base types" algorithm, one in
foundation-record.cpp and one in foundation-typeinfo.cpp.

This patch adds a __MCRecordTypeInfoGetFieldCount() function, which
assumes that the typeinfo it is passed has already been resolved.

The implemented algorithm is iterative rather than recursive, for two reasons:

1) In the original __count_fields() implementation, the recursive call
   was not in a tail position; optimisation to a loop would have been
   non-trivial, and therefore an explicit loop is more efficient.

2) Infinite loops are easier to debug than stack overflows.