[[ Cleanup ]] libfoundation: Add CODEPOINT_NONE constant for codepoint_t values.
libfoundation defines constants for limits of most basic types, but
none for codepoints.  In the engine source, there are places where a
codepoint_t variable must be checked for or assigned an value that
corresponds to no valid Unicode codepoint.  Often the value -1 is
used, but this triggers warnings.

This patch adds a CODEPOINT_NONE constant that can be used instead of
-1 in these situations.