[[ Cleanup ]] MCExternalVariable: Avoid double-dereferencing pointers.
Previously, MCExternalVariable's AppendInteger() and AppendCardinal()
methods each cast their 32-bit integer argument to a pointer and
dereferenced it, which was clearly the Wrong Thing on any non-32-bit
platform.

When I looked into this further, I found that both call sites for
these methods passed in values, not pointers, so the previous
behaviour was Obviously Wrong.

If this change causes regressions then we'll need to fix the API to be
64-bit safe and a lot less abusive.