[[ Build ]] Linux: Pass -rdynamic when linking lc-test.
The MLC runtime uses dlsym(3) to look up symbols in the main program.
Normally, GNU ld(1) does not include a dynamic symbol table in a
statically linked program.  Passing the -rdynamic flag to gcc(1) when
linking causes the --export-dynamic flag to be passed to ld(1), which
enables generation of the dynamic symbol table.