MLC: Add "the local time" syntax.
Add a new "com.livecode.date" module with a single syntax definition:

    the local time

This evaluates to a list containing the numerical components of the
local time, except for the UTC offset.  The precision of the time is
in seconds.  For example, if the current local time was:

    2015-01-16 14:45:55 +0100

then "the local time" would evaluate to:

    [ 2015, 1, 16, 14, 45, 55 ]

In the future, this syntax will return a date/time object that is
UTC-offset-aware and that can be formatted, converted, compared, etc.