[[feature-systemsyntax]] MLC: Add `quit [with status <s>]` syntax
Add a new syntax for exiting the program, optionally providing an exit
status number.

This is called `quit` instead of `exit` because the latter keyword is
already used for something else.  Also, the behaviour of `quit` then
matches the behaviour in LiveCode script.

There is an outstanding issue with this feature: `quit` just calls
exit(3) directly, which means that the "shutdownRequest" message
behaviour of LiveCode script's `quit` syntax isn't implemented.