[[CID 16909 etc.]] Avoid overruns of hash table capacities array.
Previously, hash table size lookups were performed by looping over up
to 64 entries of the __kMCValueHashTableCapacities array.
Unfortunately the array contains only 40 entries, and requesting a
size between 0x1c2ffdf3 and 0xffffffff (UINDEX_MAX) would result in a
overrunning the array.

Instead of using a fixed iteration count, add UINDEX_MAX as a
custodian to the end of the hash table capacities array and iterate
until it is reached.

Coverity-ID: 16909
Coverity-ID: 16790
Coverity-ID: 16722
Coverity-ID: 16245