[[ ScriptifyLibURL ]] Move revliburl from a button script to a script-only stack in ide-support.
This change is to make it easier to modify and update liburl. It moves the
script of button 'revliburl' of stack 'revlibrary' into a script-only stack
in the ide-support folder.

This commit goes along with r1369 and r1370 in the svn IDE repo which contain
the necessary changes to revLibrary and Home to ensure everything works as it
did before.

The version of the liburl script here is taken from the head of trunk in the
SVN repo - which I believe is the latest shipping version.