[[ LCB ]] Improved foreign types.
The 'builtin' foreign types such as bool are no longer builtin. They are now defined in the module com.livecode.foreign (in libstdscript) as 'foreign types'.

The main ones are part of libfoundation, but more specific ones should be defined in module-foreign.cpp - a NativeCString type has been added as an example.

The compiler will no longer complain about using a 'foreign' type in high-level context - it will work, but will be quite inefficient as foreign types which have a 'bridge' specify always convert to/from the bridge type when moving to/from slots unless they are an argument to a foreign function.