[[ Bug 14422 ]] Allow specification of min_os_version fields when building mach-o standalones.
The plist in an app bundle must have a MinOsVersion field which matches the mach-o load command inside the executable.

This patch adds a 'min_os_version' key to the deploy array used when using the deploy command. The key's value can either be a string or an array.

In the former case, the string is taken to be a version to apply to all architectures in the (fat) binary.

In the latter case, the array is expected to be a mapping from architecture string (key) to version (value). The architecture string can be one of:
  i386, x86_64, armv6, armv7, armv7s, arm64, ppc, ppc64
If the architecture string is empty it is taken to be 'unknown'.

The version (value) must be in the X.Y.Z where X, Y, Z are integers.

For example:
   put "6.1.0" into tDeploy["min_os_version"]["i386"]
   put "8.1.0" into tDeploy["min_os_version"]["armv7"]
   put "7.1.0" into tDeploy["min_os_version"]["arm64"]

When building a standalone, the deploy command will rewrite the min os version load command in each slice based on the architecture of the slice. If a given architecture is not present in the min_os_version array, then it will take the value specified for the 'unknown' (empty key). If neither a given architecture is present nor an 'unknown' entry then the deploy command will not touch the min_os_version load command.