[[ Build ]] Quieten non-Make build actions on Linux with 'make -s'
In some of the Linux Makefiles, there are rules that use a loop to
construct a bunch of commands to perform (for example, for compiling
MLC files).  These rules print the command that's being run for
debugging purposes, but it makes the make output unnecessarily noisy
when '-s' is specified.

This patch makes those rules obey 'make -s'.