[[ LCB Compiler ]] Generalize 'metadata'.
The metadata clause has been generalized to a definition so that it can appear anywhere at top-level.

This allows metadata to be used to provide information to the host environment (via the manifest file) - for example, extra information about properties for the inspector to use.

The metadata clause can now take a literal string as the 'key' and thus allow any character to be used.