[[ LCB Compiler ]] Generalized 'metadata' clause.
The 'metadata' clause has been promoted to the level of a definition, meaning it can appear anywhere in a module after the 'use' clauses'.

The syntax of the metadata clause is now as follows:

   metadata NameOrString is String

The compiler does nothing with metadata clauses apart from pass them through to the manifest XML file.

Metadata clauses now appear as follows in the XML file:

  <metadata key="KEY">VALUE</metadata>

Note that a warning will be generated if metadata clauses are present before use clauses. Existing code should be updated to stop the warning.