[[ LCB Compiler ]] Enable ability to specify unreserved keywords.
By default any keywords specified in syntax clauses will be reserved by the compiler. This means that they cannot be used as identifiers.

So that we can choose to unreserve keywords which we know will not cause a problem you can now specify a keyword in the form "..."! - the postfix ! means that the keyword will be unreserved.

Note: At the moment only one instance of a keyword needs to be marked as unreserved for it to be so.