[[ Bug 14485 ]] Errors in parsing expressions involving widget.
The MCChunk object is highly dependent on the order of CT_* enum for controls. In particular, CT_FIELD must be the last object type as there is a comparison which depends on it being a text chunk. Therefore CT_FIELD has been moved to after CT_WIDGET and aliased with CT_FIRST_TEXT_CHUNK. Places in the code which were comparing with CT_FIELD to get a control have been updated to CT_LAST_CONTROL, and places which were comparing to CT_FIELD as a text chunk have been updated to use CT_FIRST_TEXT_CHUNK.

In particular, this fixes the issue with parsing chunk lists such as 'widget 1 and button 2' - which was parsing 'widget (1 and button 2)' rather than 'widget (1) and button (2)'.