com.livecode.unittest: Change syntax for "expected fail" tests
Consider a test that UTF-8 decoding fails for bad input:

    test decode(tBad) is not defined

Now suppose that it's not yet implemented properly.  In the previous syntax:

    test fails when decode(tBad) is not defined

This gave the mistaken impression that the correct outcome would be
"decode(tBad) is defined".

Instead, this patch implements a new syntax:

    broken test decode(tBad) is not defined

"broken" is more easily searchable, and having the "expected fail"
indicator at the start of the line makes it easier to find.