[[ Foundation ]] Improved auto value ref classes.
New methods In, Out and InOut have been added to the AutoValueRef class. These should be used to pass the auto-container to arguments of those specific modes.

New methods Give and Take have been added to the AutoValueRef class. Give should be used to move a value into the container, and Take should be used to take a value out of the container.

There are now immutable and mutable variants of the AutoValueRef base. The immutable variants are for valuerefs which are never mutable - these have remained the same. The mutable variants have extra methods MakeMutable and MakeImmutable. These transform the mutability of the contained valueref in-place.