[[ Bug 14504 ]] Make sure strings that can be native have the kMCStringFlagCanBeNative flag set.
In particular strings which have been 'unnativized' by the engine were losing the flag when modified, due to the non-transference of flags in low-level string copying functions.
This also removes the need for a workaround associated with bug 12676, which was in turn causing slowness of native regex matching reported in bug 14481.