[[ Bug 14514 ]] Ensure that chained imports resolve properly.
It is now possible for an import definition to reference an external definition in another module. This means that resolution of a definition needs to be changed. This has been implemented by ensuring a module is valid before importing from it, and then using the resolved imports in that module, when resolving an external definition whilst the using module is being made usable.