[[ Bug 14532 ]] Prevent excessive substring copying when resolving text chunks.
Every time part of a text chunk expression is resolved, the previously calculated code unit indices are used to create a substring. This also happens at the beginning of text chunk resolution. Therefore every text chunk resolution involves at least one unnecessary copy.

This would be much better if parts of text chunk expressions modified the range accordingly, thereby preventing any copying of substrings of the underlying string until the final evaluation (or potentially not at all if this is a 'put into')