[[ LCB ]] Added trial support for context variables.
A context variable is a module-scope variable which notionally resets to its default value on every new script frame.

To declare a context variable use:

context variable zFoo [ as FooType ] default DefaultValue

The current implementation uses one flat vector for all context variables meaning they should be used sparingly. Each frame can have a list of context slots, but these are only initialized when a context variable is mutated in the context of the frame.