[[ WIP ]] Implemented real-only variants of number operations. [[ WIP ]] Changed things so that integer < number and real < number but integer is not related to real.
The arithmetic in LCB now works as follows:

There are two concrete numeric types:
  Integer - which can currently hold an integral value in the range [INT32_MIN..UINT32_MAX]
  Real - which can currently hold any IEEE double value

Both of these types are implemented as variants of MCNumberRef.

Integer literals are considered to be of type 'Integer' (this is any sequence of digits without '.' or exponent).
Real literals are considered to be of type 'Real' (real literals must include a '.' or an exponent).

If you perform an operation on two Integers then you will get Integer semantics - in particular, numeric overflow will be detected and an error raised.

If you perform an operation on two Reals then you get Real semantics - in particular, numbers are mostly approximately represented.

If you perform an operation between an Integer and a Real then the Integer is promoted to a Real and you get Real semantics.

Either an Integer or a Real can be assigned to a slot of type 'number', but an Integer cannot be assigned to a slot of type 'real' and a Real cannot be assigned to a slot of type 'integer'.

Syntax needs to be added to allow explicit conversion between real and integer:
  - <number> as real: approximate <number> by an IEEE double
  - <number> as integer: take the integer part of floor(<number), throwing an overflow error if the resulting value will not fit in Integer's permissable range.