[Bug 14482] Added a new "solid" status bar style for iOS
'solid' style is opaque and automatically shifts down the app view by 20 pixels (i.e the height of the status bar)
Note that this patch is paired with an update in the IDE. More specifically, on stack:
Toolset/revstandalonesettings.rev --> card iOS --> button Status Bar Style --> Property Inspector
Add a new menuitem "Solid". This update is not yet pushed. I will push as soon as this pull request is reviewed.