[[ BehaviorStacks ]] Behaviors generalized to stacks and buttons.
The behavior property can now be set to be either the script of a button, or the script of a stack.

This is file-format-backwards-compatible change - referencing a stack will cause the behavior reference to be written out to the stackfile with id 0 which an accessible runtime object can never have an id of.

When fetching the behavior property it might now either be a string of the form:

    button id <id> of stack <stack>

or

    stack <stack>

This might have implications on script which processes the behavior property directly, rather than relying on chunk parsing.

To extend behaviors to be able to reference any object we will need to extend the chunk syntax to allow

    object id <id> of stack <stack>

This is because ids can uniquely identify cards, audio clips, video clips as well as controls.