libsystem: Always use standard streams in unbuffered mode.
Currently the standard streams are only used for diagnostics output to
the terminal.  The default block-buffered mode for stdio is causing
issues for applications such as the testsuite.  It's currently
impossible to get at the underlying FILE pointers for stderr and
stdout because LCB can't bind foreign variables yet.

Running in unbuffered mode will slightly reduce performance for
applications that make heavy use of the standard streams for I/O.  If
this turns out to be a serious problem, we should add an API for
controlling the buffering in a LCB-friendly way.