[[ Bug 13110 ]] Debugger triggers when using 'do'.
The debugger will trigger a breakpoint on the first few lines of a script when a do is executed in the context of the same object.

Essentially, the check for breakpoints gets confused in this case so if you do script has 3 lines, and you have a breakpoint in the first 3 lines of the objects script, you'll get a debug-break action.

This has been fixed by guarding MCB_trace against doing anything if within a handler called 'message' - which is what do uses for this purpose.