LCB: Rename standard library types.
The current standard library uses lower-case type names.  This has
some unfortunate interactions with the current syntax implementation;
for example, the use of an "index" type means that it's not possible
to add any syntax that uses "index" as a keyword.

Initial draft of spec for new naming convention for standard library
types.