[[ LCB Engine ]] Changed 'dispatch' into 'post' and 'send.
The dispatch syntax has been removed and has been replaced by 'post' and 'send'.

The 'send' form is the same as dispatch.

The 'post' form will post a message to the event queue which will resolve at the next wait point. It is re-entrant safe for widgets to use and thus is more suitable for event sending (assuming you don't need a result of execution of the message!).