[[ Bug 14783 ]] Check signatures of property getters and setters.
The compiler will now check that the signatures of property getters and setters is correct.

In particular, it will ensure the 'get' handler takes no arguments and the 'set' handler will take a single 'in' argument.

The return value of the set handler is not checked since that would violate 'untyped' laxness rules.