[[ LCB ]] Detect too big integer literals.
The compiler will now detect and report an error if an integer literal is too big to fit into an unsigned 32-bit integer.

The internal literal representation has been changed to unsigned so that the full range of values can be achieved (literals are never -ve).

Note: The standard gentle value type is 'long', but integer literals only enter at MakeIntegerLiteral, and leave at EmitUnsignedIntegerConstant.