[[ LCB ]] Tame 'lc-bootstrap-compile' arguments
The 'lc-bootstrap-compile' command now takes the following arguments when in '--bootstrap' mode:

* --inputg: the input template grammar file to use
* --outputg: the output grammar file to generate
* --outputc: the output C file containing the binary modules
* --outputi: the output folder to place all the module interface files into

All other (file) arguments are taken to be the input files to process.

The bootstrap phase now processes all modules together and the complete set of dependencies for all the modules must be presented together on the command-line.
The compiler will sort out compilation order, so they don't have to be in dependency order.

All the interface files for all the specified modules will be placed into the specified 'outputi' folder.

All the binary data for the compiled modules will be emitted as C arrays into the single 'outputc' file.