[[ Bug 14993 ]] Change 'undefined' to 'nothing'
Using the term 'undefined' to describe 'no value' in LCB was confusing and so it has been replaced by 'nothing'.

The 'nothing' keyword can be used as an expression which evaluates to 'no value'.

The 'nothing' keyword can be used as a type in which case it expresses the 'no value type' which has a single value of 'nothing'.

In this context, 'optional' types can be seen as union(nothing, Type) and they are automatically initialized to 'nothing'.

Two optional values can now also be compared using 'is' and 'is not' in the way one would expect. i.e. If both values contain nothing then X is Y returns true.

The notion of whether a variable is 'defined' or not is then one which is not exposed to runtime aspects of the language beyond attempting to fetch a variable which has yet to have a value put into it.