gyp: Specify iOS SDK when running gyp.
Tools such as perfect need to be compiled against the host (Mac OS
X) SDK, even when compiling for iOS or iPhoneSimulator.  Passing
'-sdk' to xcodebuild is too much of a blunt instrument.  Instead,
configure the target SDK on the gyp command line with '-Dhost_sdk='
and/or '-Dtarget_sdk='.

For Mac OS X, the default target and host SDK is 'macosx'.  For
iOs, the default target SDK is 'iphoneos' and the default host SDK
is 'macosx'.