[[ CmdLineBuilder ]] Installers can now be built from the command-line.
This patch contains minor changes to the engine to allow the community IDE engine to run in -ui along with various changes to the builder system to enable installers to be built from the command line.

Specifically, there is a new stack 'builder_tool.livecodescript' which gives a command-line interface to the installer builder system:

<ide-engine> -ui builder/builder_tool.livecodescript

This takes the following arguments:

  . --platform <plat> : <plat> is one of win, mac, linux (multiple --platform arguments are allowed)
  . --stage <stage> : <stage> is one of environment, tools, server, notes, docs
  . --edition <edition> : <edition> is one of community, commercial
  . --build <build> : <build> is one of stable, maintenance, development, beta
  . --engine-dir <dir> : <dir> is a folder containing zipped binary components
  . --output-dir <dir> : <dir> is the folder where the final built installers should be placed
  . --work-dir <dir> : <dir> is the folder where intermediate files should be placed
  . --warn-as-error : if specified, the build will stop at the first warning

The environment stage will generate a suitable public/private key.

The docs stage will generate the documentation.

The notes stage will generate the release notes.

The tools stage will generate the installers themselves.