Add a short delay after launching mspdbsrv
As far as can be determined, our build server is too fast (!) and
starts msbuild before mspdbsrv is ready. To avoid this, we need a
short delay before starting it.

Unfortunately, there's no build-in batch command for sleeping so
we do the next-best thing and run a command with an internal
delay; ping is as good as anything else...