[[ Bug 14466 ]] 'read from socket until empty' doesn't necessarily return pending data
The MCS_read_socket() primitive will now MCS_handle_sockets() before it processes socket data to ensure that the socket state is in sync with the system socket state.

On Windows, MCS_handle_sockets() forces all WM_USER events sent to sockethwnd to be dispatched; whilst on other platforms it forces a 0ms select on all the socket handles.