[[ Bug 15408 ]] Deleted controls can still be referenced by id.
When a control was deleted by script, only the id cache for that object was being flushed leaving any child objects in the id cache.

This has been fixed by making 'scheduledelete' virtual with a parameter to determine whether it is for a child, or direct. This then recurses to scheduledelete each child.