build: Upload to server with subtly changed permissions.
On the buildserver, the umask for the build is set to 0077,
meaning that the installers end up with 0700 permissions (etc.)

This means that on upload to the distribution server, the
permissions get set wrong when --archive is passed to rsync.

Instead, set rsync to use the target umask, preserving
executability.