[[ IsReally ]] New operators 'is [not] really'.
The following operators have been added:

  x is really nothing
  x is not really nothing
  x is really a boolean
  x is not really a boolean
  x is really an integer
  x is not really an integer
  x is really a real
  x is not really a real
  x is really a string
  x is not really a string
  x is really a binary string
  x is not really a binary string
  x is really an array
  x is not really an array

These test the actual type of the value of x to see if it exactly matches the specified type.

Note that this is mainly added to help work out what types are used in values coming from widgets (in particular, widget object arrays) and due to the way the engine currently works there might be some surprising things they are not intended for general purpose use.