[[ ObjectGC ]] More efficient object deletion.
The engine will now delete all objects it can after each statement execution.

When an object is created, it is tied to a deletion pool owned by the current
wait loop. If the object is subsequently deleted (through script) whilst that
loop is still running, it will be destroyed after the current command finishes
executing.

If an object survives the current wait loop and subsequently deleted it will
be placed in the most suitable wait loop which still exists, and get deleted
when that is returned to.