[[ Actions ]] Unified redraw and drain actions.
There is now a global concept of 'Actions'. These are things that might need to run after each
command execution.

The decision on whether to run the actions depends on the bitmask set in a global variable
meaning that the action codepaths are only hit if something has requested they should be.