[[ ForeignCallbacks ]] Implement basic version of C callable LCB closures.
Code to compute the FFI layout type for handler types has been moved into MCHandlerTypeInfo.

Additionally, an MCHandlerGetFunctionPtr API has been added which attempts to create and then
return a C function pointer which will execute the HandlerRef.

In order to ensure that the Handler value and its closure are not released it is important to
keep the handler value in a variable for the lifetime the function pointer is required.