[[ Bug ]] Syntax bindings not working correctly in lc-compile.
When choosing the list of possible methods to bind to for a particular statement, lc-compile
will cut out all canvas and widget possibilities unless those modules are explicitly used.

However, rather than check this for each possible binding, it was only checking whether the
first binding was in canvas or widget thus making an implicit (incorrect) dependency on the
order of modules passed to the syntax compiler.