[[ LCB ]] New '--deps' option for lc-compile.
The lc-compile command now has a '--deps' option to generate dependency information.

The first form is to specify either '--deps order' or '--deps changed-order'. In this mode
lc-compile will read in all input files specified and then output them in an ordered list
which is needed to compile them appropriately based on dependencies between them. If
'changed-order' is used then only files which need to recompiled based on timestamps
(comparing with the interface files). Note that you need to specify a --modulepath for
these modes to work, and lc-compile will not generate errors if it cannot find specific
interface files.

The second form is to specify '--deps make'. In this mode if you specify an --output option
it will generate makefile compatible dependency rules. In this case you must only specify
a single input file.

The third form is to specify '--deps make' without an output option. In this case, a Makefile
compatible list of dependencies will be generated specifying the dependencies between the
input source files and the interface files.