[[ ComposedWidgets ]] Initial implementation of composed widgets.
The intrinsic part of a widget has been refactored into a MCWidgetRef valueref.

There are two variants of MCWidgetRef - one is a 'host' widget which is created when a widget is directly
embedded in a stack as an MCWidget; the other is a 'child' widget which is created when a widget is
used as a child widget within another widget.

There is new widget syntax to create, place, unplace and manipulate child widgets.