[emscripten] Start implementing HTML5 event handling.
Initial ability to register to receive HTML5 events, and to handle
them in the main loop.  On running an event handler, the handler:

* Adds a closure around the event structure and schedules it using
  LiveCodeAsync.delay().

* Calls LiveCodeAsync.resume().

The event handling closure uses the MCEventQueue API to post the
incoming event for handling.  This means that when the main loop
resumes, it has some new events queued to be dispatched as necessary.