[Bug 15581] libgraphics: Draw Pango layout texts into full-size buffer.
Newer versions of Pango appear to cope badly with drawing into buffers
that are smaller than the full extents of the layout to be drawn.
Specifically, a buffer overrun can be reliably triggered by attempting
to draw glyphs into a 1 px-wide FT_Bitmap pixel buffer.

This patch modifies text rendering on Linux to always allocate a
full-size buffer for Pango to render into, but only draw the clipped
area to the target surface.