libfoundation: Add new MC_DLLEXPORT_DEF macro.
The `used` attribute is needed for Emscripten builds to ensure that
native functions accessed by name via dlsym() and/or from JavaScript
keep their names and are not removed even when aggressive optimisation
is used.

`used` can only be attached to definitions, so it cannot be added to
`MC_DLLEXPORT`, because that is attached to declarations.  It is also
not possible to move `MC_DLLEXPORT` from declarations to definitions,
because on Windows, the `dllexport` attribute must be attached to
declarations (it changes the storage type of the symbol to which it is
attached).

This patch therefore adds a new `MC_DLLEXPORT_DEF` macro that must be
attached to the definitions of exported functions or variables, while
`MC_DLLEXPORT` is reserved for advance declarations (e.g. in header
files).

All symbols that are declared `MC_DLLEXPORT` are now defined with
`MC_DLLEXPORT_DEF`, and all uses of `MC_DLLEXPORT` on definitions are
replaced with `MC_DLLEXPORT_DEF`.